Correction erreurs de compilation visual studio 8.

+++ Testing/Code/BasicFilters/otbLogPolarTransformResample.cxx
@@ -57,8 +57,8 @@ try
     params[0]=0.5*static_cast<double>(reader->GetOutput()->GetLargestPossibleRegion().GetSize()[0]);
     params[1]=0.5*static_cast<double>(reader->GetOutput()->GetLargestPossibleRegion().GetSize()[1]);
     params[2]=360./1024;
-    params[3]=log(sqrt(pow(reader->GetOutput()->GetLargestPossibleRegion().GetSize()[0],2)
-+pow(reader->GetOutput()->GetLargestPossibleRegion().GetSize()[1],2))/2)/512;
+    params[3]=log(sqrt(pow(static_cast<double>(reader->GetOutput()->GetLargestPossibleRegion().GetSize()[0]),2)
++pow(static_cast<double>(reader->GetOutput()->GetLargestPossibleRegion().GetSize()[1]),2))/2)/512;
     transform->SetParameters(params);
 
     ImageType::SizeType size;
